requirements.txt
langchain 
langchain-community 
langchain-google-genai 
unstructured[all-docs] 
pillow_heif 
pdf2image
pytesseract
poppler-utils
langdetect
unstructured-inference
unstructured-pytesseract
chromadb
python-dotenv
langchain
langchain-core
sentence-transformers
streamlit
langchain-huggingface